Big refactoring going from nested classes to all separate ones
