+++ languageServer/src/main/scala/core/bigrammar/grammars/Delimiter.scala
@@ -6,5 +6,5 @@ case class Delimiter(value: String) extends StringGrammar {
   if (value.length == 0)
     throw new RuntimeException("value must have non-zero length")
 
-  override def getParser(keywords: scala.collection.Set[String]): BiGrammarToParser.Parser[String] = BiGrammarToParser.Literal(value)
+  override def getParser(keywords: scala.collection.Set[String]): BiGrammarToParser.Parser[String] = BiGrammarToParser.literal(value)
 }
