+++ languageServer/src/main/scala/core/bigrammar/grammars/Keyword.scala
@@ -7,6 +7,5 @@ case class Keyword(value: String, reserved: Boolean = true, verifyWhenPrinting: 
   if (value.length == 0)
     throw new RuntimeException("value must have non-zero length")
 
-  override def getParser(keywords: scala.collection.Set[String]): BiGrammarToParser.Parser[String] =
-    BiGrammarToParser.Literal(value)
+  override def getParser(keywords: scala.collection.Set[String]): BiGrammarToParser.Parser[String] = BiGrammarToParser.literal(value)
 }
