+++ languageServer/src/main/scala/core/parsers/Fail.scala
@@ -1,0 +1,7 @@
+package core.parsers
+
+case class Fail[Input <: ParseInput](message: String) extends Parser[Input, Nothing] {
+  override def parse(input: Input, cache: ParseState): ParseResult[Nothing] = ParseFailure(None, input, message)
+
+  override def getDefault(cache: DefaultCache): Option[Nothing] = None
+}
