+++ languageServer/src/main/scala/core/parsers/MapParser.scala
@@ -1,0 +1,9 @@
+package core.parsers
+
+class MapParser[Input <: ParseInput, +Result, NewResult](original: Parser[Input, Result], f: Result => NewResult) extends Parser[Input, NewResult] {
+  override def parse(input: Input, cache: ParseState): ParseResult[NewResult] = {
+    original.parseCached(input, cache).map(f)
+  }
+
+  override def getDefault(cache: DefaultCache): Option[NewResult] = cache(original).map(f)
+}
